requirements.txt
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
appscript==1.2.5
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bitstruct==8.11.1
build==1.2.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
contourpy==1.0.7
coverage==7.6.3
crashtest==0.4.1
cssselect==1.2.0
cycler==0.11.0
decorator==4.4.2
distlib==0.3.8
distro==1.9.0
Django==5.1.3
doc2docx==0.2.4
docstring_parser==0.16
dulwich==0.21.7
EbookLib==0.18
et-xmlfile==1.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fit-tool==0.9.13
fonttools==4.40.0
fsspec==2024.9.0
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.20.0
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
html-from-epub==0.2.1
htmldocx==0.0.6
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.25.1
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jdcal==1.4.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.2.0
keyring==24.3.1
kiwisolver==1.4.4
lxml==5.3.0
lxml-stubs==0.5.1
Markdown==3.7
markdown2==2.5.0
markdownify==0.13.1
MarkupSafe==3.0.2
matplotlib==3.7.1
more-itertools==10.5.0
moviepy==1.0.3
msgpack==1.1.0
mypy==1.11.2
mypy-extensions==1.0.0
numpy==1.24.3
oauthlib==3.2.2
openai==1.44.0
openpyxl==2.5.12
packaging==23.1
pandas==2.0.2
pexpect==4.9.0
Pillow==9.5.0
pkginfo==1.11.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==4.25.4
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.0.9
pyproject_hooks==1.1.0
pytest==7.0.1
pytest-cov==3.0.0
pytest-mock==3.6.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tenacity==9.0.0
threadpoolctl==3.1.0
tokenizers==0.20.0
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.66.5
trove-classifiers==2024.9.12
types-beautifulsoup4==4.12.0.20240511
types-html5lib==1.1.11.20240806
types-lxml==2024.8.7
types-Pillow==10.2.0.20240822
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2023.3
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.5
xattr==1.1.0
zipp==3.20.2
